Show gcode preview image before print if available
To include PrusaSlicer thumbnails:
Make a copy of the the printer profile, and go to menu>help>Show configuration folder.
Locate the printers subfolder and edit the .ini file of your printer.
Around the bottom (line 69), there is a line that says "thumbnails =" change it to "thumbnails = 16x16,220x124".
Then add gcode_thumb_to_jpg.py to the to the post-processing scripts section of print settings.